requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
Flask==3.1.0
Flask-Session==0.8.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
marshmallow==3.25.1
msgspec==0.19.0
mypy-extensions==1.0.0
nylas==6.5.0
packaging==24.2
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
gunicorn